update: change ponderQueries to usePonder and remove fetch data from status offer
